move HasAccess off DynamicCache
