+++ internal/objectstore/dynamic_cache_test.go
@@ -118,7 +118,9 @@ func Test_DynamicCache_List(t *testing.T) {
 		}
 	}
 
-	c, err := NewDynamicCache(ctx, client, factoryFunc)
+	resourceAccess := NewResourceAccess(client)
+
+	c, err := NewDynamicCache(ctx, client, factoryFunc, Access(resourceAccess))
 	require.NoError(t, err)
 
 	key := store.Key{
@@ -187,7 +189,8 @@ func Test_DynamicCache_Get(t *testing.T) {
 		}
 	}
 
-	c, err := NewDynamicCache(ctx, client, factoryFunc)
+	resourceAccess := NewResourceAccess(client)
+	c, err := NewDynamicCache(ctx, client, factoryFunc, Access(resourceAccess))
 	require.NoError(t, err)
 
 	key := store.Key{
@@ -204,120 +207,6 @@ func Test_DynamicCache_Get(t *testing.T) {
 	expected := testutil.ToUnstructured(t, pod)
 
 	assert.Equal(t, expected, got)
-}
-
-func Test_DynamicCache_HasAccess(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	controller := gomock.NewController(t)
-	defer controller.Finish()
-
-	client := clusterfake.NewMockClientInterface(controller)
-	informerFactory := clusterfake.NewMockDynamicSharedInformerFactory(controller)
-	namespaceClient := clusterfake.NewMockNamespaceInterface(controller)
-
-	client.EXPECT().NamespaceClient().Return(namespaceClient, nil).MaxTimes(3)
-	namespaces := []string{"test", ""}
-	namespaceClient.EXPECT().Names().Return(namespaces, nil).MaxTimes(3)
-
-	factoryFunc := func(c *DynamicCache) {
-		c.initFactoryFunc = func(context.Context, cluster.ClientInterface, string) (dynamicinformer.DynamicSharedInformerFactory, error) {
-			return informerFactory, nil
-		}
-	}
-
-	scenarios := []struct {
-		name       string
-		resource   string
-		key        store.Key
-		accessFunc func(c *DynamicCache)
-		expectErr  bool
-	}{
-		{
-			name:     "pods",
-			resource: "pods",
-			key: store.Key{
-				APIVersion: "apps/v1",
-				Kind:       "Pod",
-			},
-			accessFunc: func(c *DynamicCache) {
-				ac := initAccessCache()
-				aKey := accessKey{
-					Namespace: "",
-					Group:     "apps",
-					Resource:  "pods",
-					Verb:      "get",
-				}
-				ac.access[aKey] = true
-				c.access = ac
-			},
-			expectErr: false,
-		},
-		{
-			name:     "crds",
-			resource: "customresourcedefinitions",
-			key: store.Key{
-				APIVersion: "apiextensions.k8s.io/v1beta1",
-				Kind:       "CustomResourceDefinition",
-			},
-			accessFunc: func(c *DynamicCache) {
-				ac := initAccessCache()
-				aKey := accessKey{
-					Namespace: "",
-					Group:     "apiextensions.k8s.io",
-					Resource:  "customresourcedefinitions",
-					Verb:      "get",
-				}
-				ac.access[aKey] = true
-				c.access = ac
-			},
-			expectErr: false,
-		},
-		{
-			name:     "no access crds",
-			resource: "customresourcedefinitions",
-			key: store.Key{
-				APIVersion: "apiextensions.k8s.io/v1beta1",
-				Kind:       "CustomResourceDefinition",
-			},
-			accessFunc: func(c *DynamicCache) {
-				ac := initAccessCache()
-				aKey := accessKey{
-					Namespace: "",
-					Group:     "apiextensions.k8s.io",
-					Resource:  "customresourcedefinitions",
-					Verb:      "get",
-				}
-				ac.access[aKey] = false
-				c.access = ac
-			},
-			expectErr: true,
-		},
-	}
-
-	for i := range scenarios {
-		ts := scenarios[i]
-		t.Run(ts.name, func(t *testing.T) {
-			fn := ts.accessFunc
-			c, err := NewDynamicCache(ctx, client, factoryFunc, fn)
-			require.NoError(t, err)
-
-			gvk := ts.key.GroupVersionKind()
-			podGVR := schema.GroupVersionResource{
-				Group:    gvk.Group,
-				Version:  gvk.Version,
-				Resource: ts.resource,
-			}
-			client.EXPECT().Resource(gomock.Eq(gvk.GroupKind())).Return(podGVR, nil)
-
-			if ts.expectErr {
-				require.Error(t, c.HasAccess(ctx, ts.key, "get"))
-			} else {
-				require.NoError(t, c.HasAccess(ctx, ts.key, "get"))
-			}
-		})
-	}
 }
 
 func TestDynamicCache_Update(t *testing.T) {
@@ -378,7 +267,8 @@ func TestDynamicCache_Update(t *testing.T) {
 
 	client.EXPECT().DynamicClient().Return(dc, nil)
 
-	c, err := NewDynamicCache(ctx, client, factoryFunc)
+	resourceAccess := NewResourceAccess(client)
+	c, err := NewDynamicCache(ctx, client, factoryFunc, Access(resourceAccess))
 	require.NoError(t, err)
 
 	key, err := store.KeyFromObject(pod)
@@ -446,7 +336,8 @@ func TestDynamicCache_Delete(t *testing.T) {
 
 	client.EXPECT().DynamicClient().Return(dc, nil)
 
-	c, err := NewDynamicCache(ctx, client, factoryFunc)
+	resourceAccess := NewResourceAccess(client)
+	c, err := NewDynamicCache(ctx, client, factoryFunc, Access(resourceAccess))
 	require.NoError(t, err)
 
 	key, err := store.KeyFromObject(pod)
